_legacy/backend/requirements.txt
google-cloud-videointelligence
google-cloud-storage
supabase
yt-dlp
python-dotenv
